requirements.txt
asgiref==3.8.1
astroid==3.3.5
black==24.10.0
cfgv==3.4.0
click==8.1.7
dill==0.3.9
distlib==0.3.9
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==30.3.0
filelock==3.16.1
identify==2.6.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
model-bakery==1.20.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
virtualenv==20.26.6
